% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fritz_log.R
\name{fritz_log}
\alias{fritz_log}
\title{Load a log file and transform into a well-formatted tibble}
\usage{
fritz_log(file)
}
\arguments{
\item{file}{A path to a log file.}
}
\value{
A tibble.
}
\description{
Load a log file and transform into a well-formatted tibble
}
\examples{
fritz_log("fritz.log")
}
