% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fritz_plot.R
\name{fritz_plot}
\alias{fritz_plot}
\title{Load a log file and create a graphical summary of device login activity}
\usage{
fritz_plot(file, plot_type)
}
\arguments{
\item{file}{A path to a log file.}

\item{plot_type}{The type of plot. Can either be "histogram" (the default) or "freq_poly".}
}
\value{
A ggplot2 plot.
}
\description{
Load a log file and create a graphical summary of device login activity
}
\examples{
fritz_plot("fritz.log", "histogram")
}
